Properly calculate root item areas (#457)
* Properly calculate root item areas

Update layout manager to properly calculate root item areas when the layout origin is not (0,0).

Currently if a dashboard is not positioned at or near 0,0 in the document, the left and top root drag areas do not work.  This fixes that by adding in the current x and y offset to the root areas when appropriate.

* Removed unused variable

* Update ES6 version as well

* Fix formatting/tab/spaces

(cherry picked from commit b118951147cc77c571fe9f4365da763c3462356e)